improve encapsulation
